Move existing dex code to stubdex component, add stubbed dex component
