+++ component/src/stubdex/stub_cpmm.rs
